Tidy-up and address review comments
