+++ extensions/ql-vscode/src/remote-queries/remote-queries-markdown-generation.ts
@@ -1,14 +1,18 @@
-import { createMarkdownRemoteFileRef } from '../pure/location-link-utils';
+import { createRemoteFileRef } from '../pure/location-link-utils';
 import { RemoteQuery } from './remote-query';
-import { AnalysisAlert, AnalysisResults } from './shared/analysis-result';
+import { AnalysisAlert, AnalysisResults, FileLink } from './shared/analysis-result';
 
 // Each array item is a line of the markdown file.
 export type MarkdownFile = string[];
 
+/**
+ * Generates markdown files with variant analysis results.
+ */
 export function generateMarkdown(query: RemoteQuery, analysesResults: AnalysisResults[]): MarkdownFile[] {
   const files: MarkdownFile[] = [];
   for (const analysisResult of analysesResults) {
     if (analysisResult.interpretedResults.length === 0) {
+      // TODO: We'll add support for non-interpreted results later.
       continue;
     }
     const lines = [
@@ -31,17 +35,15 @@ function generateMarkdownForInterpretedResult(interpretedResult: AnalysisAlert, 
     interpretedResult.fileLink,
     interpretedResult.highlightedRegion?.startLine,
     interpretedResult.highlightedRegion?.endLine
-  ), '');
+  ));
+  lines.push('');
   const codeSnippet = interpretedResult.codeSnippet?.text;
   if (codeSnippet) {
     lines.push(
-      `\`\`\`${language}`,
-      ...codeSnippet.split('\n'),
-      '```',
-      ''
+      ...generateMarkdownForCodeSnippet(codeSnippet, language),
     );
   }
-  const alertMessage = buildAlertMessage(interpretedResult);
+  const alertMessage = buildMarkdownAlertMessage(interpretedResult);
   lines.push(alertMessage);
 
   // Padding between results
@@ -53,7 +55,18 @@ function generateMarkdownForInterpretedResult(interpretedResult: AnalysisAlert, 
   return lines;
 }
 
-function buildAlertMessage(interpretedResult: AnalysisAlert): string {
+function generateMarkdownForCodeSnippet(codeSnippet: string, language: string): MarkdownFile {
+  const lines: MarkdownFile = [];
+  lines.push(
+    `\`\`\`${language}`,
+    ...codeSnippet.split('\n'),
+    '```',
+  );
+  lines.push('');
+  return lines;
+}
+
+function buildMarkdownAlertMessage(interpretedResult: AnalysisAlert): string {
   let alertMessage = '';
   for (const token of interpretedResult.message.tokens) {
     if (token.t === 'text') {
@@ -67,5 +80,20 @@ function buildAlertMessage(interpretedResult: AnalysisAlert): string {
       );
     }
   }
-  return alertMessage;
+  // Italicize the alert message
+  return `*${alertMessage}*`;
+}
+
+/**
+ * Creates a markdown link to a remote file.
+ * If the "link text" is not provided, we use the file path.
+ */
+export function createMarkdownRemoteFileRef(
+  fileLink: FileLink,
+  startLine?: number,
+  endLine?: number,
+  linkText?: string,
+): string {
+  const markdownLink = `[${linkText || fileLink.filePath}](${createRemoteFileRef(fileLink, startLine, endLine)})`;
+  return markdownLink;
 }
